+++ src/main/kotlin/com/eatsfinder/domain/like/service/PostLikeServiceImpl.kt
@@ -6,6 +6,7 @@ import com.eatsfinder.domain.like.repository.PostLikeRepository
 import com.eatsfinder.domain.post.repository.PostRepository
 import com.eatsfinder.domain.user.repository.UserRepository
 import com.eatsfinder.global.exception.ModelNotFoundException
+import com.eatsfinder.global.exception.like.DefaultZeroException
 import com.eatsfinder.global.exception.profile.MyProfileException
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
@@ -44,18 +45,28 @@ class PostLikeServiceImpl(
 
     @Transactional
     override fun deletePostLikes(userId: Long, postId: Long) {
-        val user = userRepository.findByIdAndDeletedAt(userId, null) ?: throw ModelNotFoundException("user", "이 유저 아이디(${userId})는 존재하지 않습니다.")
-        val post = postRepository.findByIdOrNull(postId) ?: throw ModelNotFoundException("post", "이 게시물 아이디: (${postId})는 존재하지 않습니다.")
+        val user = userRepository.findByIdAndDeletedAt(userId, null) ?: throw ModelNotFoundException(
+            "user",
+            "이 유저 아이디(${userId})는 존재하지 않습니다."
+        )
+        val post = postRepository.findByIdOrNull(postId) ?: throw ModelNotFoundException(
+            "post",
+            "이 게시물 아이디: (${postId})는 존재하지 않습니다."
+        )
         val postLike = postLikeRepository.findByUserIdAndPostId(user, post)
 
-        if (post.userId.id == user.id){
+        if (post.userId.id == user.id) {
             throw MyProfileException("본인 게시물이므로 좋아요를 취소할 수 없습니다.")
         }
 
-        if (postLike != null){
-            postLikeRepository.delete(postLike)
-            post.likeCount--
-            postRepository.save(post)
+        if (postLike != null) {
+            if (post.likeCount > 0) {
+                postLikeRepository.delete(postLike)
+                post.likeCount--
+                postRepository.save(post)
+            } else {
+                throw DefaultZeroException("좋아요 수의 기본값은 0입니다.")
+            }
         } else {
             throw ModelNotFoundException("like", "좋아요(${postId})는 존재하지 않아 취소할 수 없습니다.")
         }
